fix swiftlint errors
